requirements.txt
streamlit
google-generativeai
gTTS
pillow